On branch master

+++ src/main/java/edu/mum/coffee/repository/ProductRepository.java
@@ -12,19 +12,20 @@ import edu.mum.coffee.domain.Product;
 import edu.mum.coffee.domain.ProductType;
 
 @Repository
-public interface ProductRepository  extends CrudRepository<Product, Long>{
+public interface ProductRepository  extends CrudRepository<Product, Integer>{
 
 	public List<Product> findByProductNameLikeOrDescriptionLikeAllIgnoreCase(String productName, String description); 
 	public List<Product> findByProductType(ProductType productType); 
 	public List<Product> findByPriceBetween(Double minPrice, Double maxPrice);
 	
 	@Query(value= "delete from Product p where p.id=:id")
-	public void delete(Long id);
+	public void delete(int id);
 	
 	 
 	@Query(value= "SELECT p from Product p WHERE p.id =:pId")
 	public Product findById(int pId);
-	
+	//@Query(value= "update Product p  set  p.productName=?1,  WHERE p.id =:pId")
+	//public Product update(int pId, Product product);
 	
     
 }
